Separate out and enhance extensions\rTESTING=unit
