+++ lib/promoted/ruby/client/log_request_builder.rb
@@ -171,16 +171,6 @@ module Promoted
   end
 end
 
-class String
-   # Ruby mutation methods have the expectation to return self if a mutation occurred, nil otherwise. (see http://www.ruby-doc.org/core-1.9.3/String.html#method-i-gsub-21)
-   def to_underscore!
-     gsub!(/(.)([A-Z])/,'\1_\2')
-     downcase!
-   end
-
-   def to_underscore
-     dup.tap { |s| s.to_underscore! }
-   end
-end
 require 'securerandom'
 require "promoted/ruby/client/constants"
+require "promoted/ruby/client/extensions"
